[BUGFIX] Do not iterate over _processed_ folder
Instead of iterating over the _processed_ folder and
after that filtering the result, we should instead
prevent iterating over it.

This improves performance of search in root directory
by a lot, especially when the processed folder contains
a lot of files.

Also exclude the folder, when recursively iterating
over folders, because this never happens when the
processed files are needed, but is only used on search.