[BUGFIX] Improve caching of files and folders
When iterating over files in S3 storage most time is spent
for I/O wait (the web requests to Amazon).

To improve performance reading the content of a directory should only
happen once and then be cached.

Instead of building multiple cache entries for one directory
(one with files, one with folders and one with both),
we fetch store only one entry with files and folders,
and filter the result requested from TYPO3 API.

By doing so, searching from root, will warmup the complete
cache, so that other storage listings will already be cached.

Now the caching logic is moved to a new CachedDirectoryIterator,
so we can remove the cache handling from the driver.