[FEATURE] - Update remote Cache-Control directives.
If a new file is added or the metadata indexer is triggered update the metadata of the remote S3 object. This allows you to set cache control directives like `max-age` for specific file types.

Processed files are updated by monitoring the resource storage, since the filenames contain some kind of identifier the `max-age` in the dummy configuration is set to one week.